Rename analyzers
Moved type and time analyzer to a subdirectory.
