+++ src/backend/expungeservice/expungement_analyzer/analyzers/type_analyzer.py
